Reverting nkovacs commit for some reasons
There was parallel development between his fork and the original repo,
and thus the merge would not happen anyway. So I decided to pull in his
changes into that part of the history (to still refer his work) and
implement the ideas in a different way, that would anyway fit better
with the project's ideas.